Cards!
A whole lot of stuff in this one, let's break it down.

CardData determines what a card dos.
CardTemplate is the UI prefab
Card interfaces with the UI and bridges the gap between UI and data
Deck stores CardDatas
Hand visually contains Cards.

Cards can be navigated across and selected as if they were normal buttons.
When a card is used, it is discarded back to its owner deck.