Trade Row and Player Two
The Trade Row is controlled by a Trade Deck, which deals identical TradeCards to each players Trade Row. If a TradeCard is bought, its twin in the opponents Trade Row will be destroyed. Bought Trade Row cards go to the deck of whoever bought them.

Also updated P2's UI and controls to be able to interface with the cards.